modelzoo/requirements.txt
cerebras_pytorch==2.4.2
tabulate==0.9.0
transformers==4.40.0
tokenizers==0.19.1
datasets==2.19.1
filelock==3.14.0
more-itertools==9.1.0
datasketch==1.5.9
ftfy==6.1.1
networkit==10.1
Keras-Preprocessing==1.1.2
scipy==1.10.1
regex>=2021.8.3
nltk==3.8.1
spacy==3.7.5
matplotlib==3.4.3
pyarrow==12.0.1
pyYAML<6.0
pandas==1.3.0
jsonschema==4.19.2
torch==2.4.0
torchvision==0.19.0
safetensors==0.4.1
sentencepiece==0.1.99
lm-dataformat @ https://github.com/leogao2/lm_dataformat/archive/ac85cb7dae49ce25e9973a128ebd9167deaf64dd.zip
lm-eval @ https://github.com/EleutherAI/lm-evaluation-harness/archive/refs/tags/v0.4.4.zip
bigcode_eval @ https://github.com/bigcode-project/bigcode-evaluation-harness/archive/f0b81a9d079289881bd42f509811d42fe73e58cf.zip
h5py==3.10.0
tqdm==4.66.1
Flask==2.3.3
Pillow==9.4.0
pydantic==2.8.2
cerebras_cloud_sdk>=1.0.0
click==8.1.7
termcolor==2.4.0
